The engine will get some serious changes in the next weeks and months. First I will clean up and change the system itself. So you will have a rendertarget. That means, you are able to render to another SDL_Surface instead of the Screen. Furthermore I will make a cut between the crossplattform handheld stuff and the engine. So you would be able to use the 3D Part of this engine e.g. with Penjin. Also I gave this engine a better name (I think): Sparrow3D. I like sparrows. ^^
Later Sparrow3D will get more features. The old sorting based drawing
will be history. You will be able to use a real z-Buffer and a simple
form of texture mapping. Furthermore I think about Alphablending. I
noticed, that the matrix operations are the slowest part of the engine,
not the drawing. So the new engine will be a bit slower, which means,
that you'll get a bit fewer polygons, but you will get textures and a
VERY better and accurate sorting.